requirements.txt
pydenticon~=0.3.1
fastapi~=0.85.1
Pillow~=9.3.0
requests~=2.28.1
gunicorn~=20.1.0
uvicorn[standard]~=0.19.0
rich~=12.6.0
rich-click~=1.5.2